Move the animation Impl classes to private headers
