+++ src/gtk/animate.cpp
@@ -13,6 +13,7 @@
 
 #if wxUSE_ANIMATIONCTRL
 #include "wx/animate.h"
+#include "wx/gtk/private/animate.h"
 
 #ifndef WX_PRECOMP
     #include "wx/image.h"
